+++ packages/2019-disaster-game/src/components/Game/TaskScreen/modals/ChosenTaskModal.js
@@ -2,7 +2,7 @@
 import { css, jsx } from "@emotion/core";
 import { connect } from "react-redux";
 import { PropTypes } from "prop-types";
-import { getActiveTask } from "../../../../state/newTasks";
+import { getActiveTask } from "../../../../state/tasks";
 import { palette } from "../../../../constants/style";
 
 const modalContainer = css`
