+++ packages/2019-disaster-game/src/state/index.js
@@ -5,7 +5,7 @@ import { kit } from "./kit";
 import { user } from "./user";
 import { settings } from "./settings";
 import { sfx } from "./sfx";
-import newTasks from "./newTasks";
+import tasks from "./tasks";
 
 export default function createReducer(asyncReducers) {
   return combineReducers({
@@ -15,7 +15,7 @@ export default function createReducer(asyncReducers) {
     user,
     settings,
     sfx,
-    newTasks,
+    tasks,
     ...asyncReducers
   });
 }
